# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Wallet API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    APIResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { type: object }
paths:
  /auth/register:
    post:
      security: []
      summary: Register a user and create the user's wallet
      responses:
        "201": { description: Created }
  /auth/login:
    post:
      security: []
      summary: Login with phone and password
      responses:
        "200": { description: OK }
  /auth/refresh:
    post:
      security: []
      summary: Rotate a refresh token and issue new credentials
      responses:
        "200": { description: OK }
  /me:
    get:
      summary: Get profile
      responses:
        "200": { description: OK }
  /me/profile-image:
    put:
      summary: Upload or update the authenticated user's profile image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
          application/json:
            schema:
              type: object
              required: [profile_image_url]
              properties:
                profile_image_url:
                  type: string
                  format: uri
      responses:
        "200": { description: OK }
  /me/kyc:
    post:
      summary: Submit KYC images and identity details
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [full_name, dob, address, id_card_image, selfie_image]
              properties:
                full_name:
                  type: string
                dob:
                  type: string
                  format: date
                address:
                  type: string
                id_card_image:
                  type: string
                  format: binary
                selfie_image:
                  type: string
                  format: binary
          application/json:
            schema:
              type: object
              required: [full_name, dob, address, id_card_image_url, selfie_image_url]
              properties:
                full_name:
                  type: string
                dob:
                  type: string
                  format: date
                address:
                  type: string
                id_card_image_url:
                  type: string
                  format: uri
                selfie_image_url:
                  type: string
                  format: uri
      responses:
        "200": { description: OK }
  /wallet:
    get:
      summary: Get wallet balance
      responses:
        "200": { description: OK }
  /wallet/deposit:
    post:
      summary: Deposit money
      responses:
        "201": { description: Created }
  /wallet/withdraw:
    post:
      summary: Withdraw money
      responses:
        "201": { description: Created }
  /wallet/transfer:
    post:
      summary: Transfer money using row-level locks and an idempotency key
      responses:
        "201": { description: Created }
  /wallet/transactions:
    get:
      summary: Cursor-paginated transaction history with type/status filters
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: type
          schema: { type: string, enum: [deposit, withdraw, transfer] }
        - in: query
          name: status
          schema: { type: string, enum: [pending, success, failed] }
      responses:
        "200": { description: OK }
  /admin/login:
    post:
      security: []
      summary: Admin login
      responses:
        "200": { description: OK }
  /admin/users:
    get:
      summary: Admin cursor-paginated list users
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200": { description: OK }
  /admin/transactions:
    get:
      summary: Admin cursor-paginated list all transactions
      parameters:
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: type
          schema: { type: string, enum: [deposit, withdraw, transfer] }
        - in: query
          name: status
          schema: { type: string, enum: [pending, success, failed] }
      responses:
        "200": { description: OK }
  /admin/users/{userID}/kyc:
    put:
      summary: Approve or reject KYC
      parameters:
        - in: path
          name: userID
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /admin/users/{userID}/status:
    put:
      summary: Freeze or unfreeze account
      parameters:
        - in: path
          name: userID
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }